feat: separar layout — docs en (docs) route group, homepage sin chrome
