+++ next.config.ts
@@ -5,6 +5,15 @@ const nextConfig: NextConfig = {
   turbopack: {
     root: path.resolve(__dirname),
   },
+  async redirects() {
+    return [
+      {
+        source: '/docs/:slug*',
+        destination: '/documentacion/:slug*',
+        permanent: true,
+      },
+    ]
+  },
 };
 
 export default nextConfig;
