+++ src/app/layout.tsx
@@ -2,10 +2,6 @@ import type { Metadata } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import './globals.css'
 import { ThemeProvider } from '@/components/ThemeProvider'
-import { Header } from '@/components/Header'
-import { Sidebar } from '@/components/Sidebar'
-import { SearchModal } from '@/components/SearchModal'
-import { getNavigation, getAllDocs } from '@/lib/docs'
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -23,13 +19,6 @@ export const metadata: Metadata = {
 }
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
-  const groups = getNavigation()
-  const searchDocs = getAllDocs().map((doc) => ({
-    title: doc.frontmatter.title,
-    description: doc.frontmatter.description ?? '',
-    href: `/docs/${doc.slug.join('/')}`,
-  }))
-
   return (
     <html
       lang="en"
@@ -38,16 +27,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
     >
       <body className="bg-white dark:bg-zinc-950 text-zinc-900 dark:text-zinc-50">
         <ThemeProvider>
-          <div className="flex h-screen overflow-hidden">
-            <Sidebar groups={groups} />
-            <div className="flex flex-1 flex-col overflow-hidden">
-              <Header />
-              <main className="flex-1 overflow-y-auto">
-                {children}
-              </main>
-            </div>
-          </div>
-          <SearchModal docs={searchDocs} />
+          {children}
         </ThemeProvider>
       </body>
     </html>
